1.1.7 - update
• Исправлена ошибка с неверным перекрашиваем клавиш клавиатуры
• Исправлена ошибка с показом игровых уведомлений в виде окошек снизу
• Исправлена ошибка с перекрашиваем текста статусного бара при смене темы
• Внесены изменения в описания некоторых достижений и условий их получения
• Скрыт ряд достижений до момента их получения! Успели запомнить?
• Добавлено 11 новых достижений!
• Ускорена работа приложения
• Исправлена ошибка с серийными достижениями

+++ app/src/main/java/com/sinya/projects/wordle/domain/model/Game.kt
@@ -9,7 +9,7 @@ data class Game(
     val mode: GameMode,
     val lang: String,
     val length: Int,
-    val totalSeconds: Long,
+    val totalSeconds: Int,
     val settings: GameSettings,
     val board: List<Cell>,
     val keyboard: List<List<Key>>,
