+++ app/src/main/java/com/sinya/projects/wordle/presentation/game/GameUiState.kt
@@ -5,7 +5,7 @@ import com.sinya.projects.wordle.domain.model.Cell
 import com.sinya.projects.wordle.domain.enums.GameMode
 import com.sinya.projects.wordle.domain.model.Key
 import com.sinya.projects.wordle.presentation.game.finishSheet.FinishStatisticGame
-import com.sinya.projects.wordle.ui.features.UiText
+import com.sinya.projects.wordle.domain.model.UiText
 
 sealed interface GameUiState {
     data object Loading : GameUiState
@@ -20,7 +20,7 @@ sealed interface GameUiState {
 
         val focusedCell: Int = 0,
         val result: GameState = GameState.NONE,
-        val timePassed: Long = 0,
+        val timePassed: Int = 0,
         val mode: GameMode = GameMode.NORMAL,
         val wordLength: Int = 5,
         val lang: String = "ru",
